requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
PyPDF2==3.0.0
python-docx==0.8.11
spacy==3.5.1
